Issue #11163: Enforced Filesize MethodLengthModifier
